+++ src/Component/Filter/TableFilter.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace srag\TableUI\Component\Filter;
+namespace ILIAS\UI\DataTable\Component\Filter;
 
-use srag\TableUI\Component\Filter\Sort\TableFilterSortField;
+use ILIAS\UI\DataTable\Component\Filter\Sort\TableFilterSortField;
 
 /**
  * Interface TableFilter
  *
- * @package srag\TableUI\Component\Filter
+ * @package ILIAS\UI\DataTable\Component\Filter
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */