+++ src/Implementation/Column/Action/ActionTableColumn.php
@@ -1,15 +1,15 @@
 <?php
 
-namespace srag\TableUI\Implementation\Column\Action;
+namespace ILIAS\UI\DataTable\Implementation\Column\Action;
 
-use srag\TableUI\Component\Column\Action\ActionTableColumn as ActionTableColumnInterface;
-use srag\TableUI\Component\Column\TableColumn as TableColumnInterface;
-use srag\TableUI\Implementation\Column\TableColumn;
+use ILIAS\UI\DataTable\Component\Column\Action\ActionTableColumn as ActionTableColumnInterface;
+use ILIAS\UI\DataTable\Component\Column\TableColumn as TableColumnInterface;
+use ILIAS\UI\DataTable\Implementation\Column\TableColumn;
 
 /**
  * Class ActionTableColumn
  *
- * @package srag\TableUI\Implementation\Column\Action
+ * @package ILIAS\UI\DataTable\Implementation\Column\Action
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */